database scanning, song deletion, tag sys updates

+++ database2.py
@@ -1,6 +1,6 @@
-import os, sys, md5, time, json, sqlite3, shutil
+import json, md5, os, shutil, sqlite3, sys, threading, time
 if 'modules' not in sys.path: sys.path.append('modules')
-from flask import Flask, Response, request, abort, make_response
+from flask import Flask, Response, request, abort
 import util, organize
 
 # Ignore database files when checking for external changes to the library's
@@ -52,7 +52,7 @@ SETUP = """
 CREATE TABLE songs (id INTEGER PRIMARY KEY, name TEXT NOT NULL,
                     track_performer TEXT,
                     album_id INTEGER, artist_id INTEGER,
-                    path TEXT, size INTEGER,
+                    path TEXT,
                     UNIQUE (name, album_id, artist_id),
                     FOREIGN KEY (artist_id) REFERENCES artists (id),
                     FOREIGN KEY (album_id) REFERENCES albums (id));
@@ -70,7 +70,7 @@ CREATE TABLE albums (id INTEGER PRIMARY KEY, name TEXT NOT NULL,
 CREATE INDEX albumnameindex ON albums (name);
 CREATE INDEX albumartistindex ON albums (artist_id);
 
-CREATE TABLE folderhashes (path TEXT PRIMARY KEY, hash TEXT);
+CREATE TABLE fsrecords (path TEXT PRIMARY KEY, size INTEGER, checked BOOLEAN);
 
 CREATE TABLE tags (id INTEGER PRIMARY KEY, name TEXT UNIQUE);
 CREATE TABLE taglinks (id INTEGER PRIMARY KEY, song_id INTEGER NOT NULL,
@@ -82,39 +82,54 @@ CREATE INDEX songlinksindex ON taglinks (song_id);
 
 CREATE TABLE globals (key TEXT PRIMARY KEY, value GLOB);
 """
-RESET = """
+CLEAR = """
 DELETE FROM songs;
 DELETE FROM artists;
 DELETE FROM albums;
-DELETE FROM folderhashes;
+DELETE FROM fsrecords;
 DELETE FROM taglinks;
+"""
+DROP_ALL = """
+DROP TABLE songs;
+DROP TABLE artists;
+DROP TABLE albums;
+DROP TABLE fsrecords;
+DROP TABLE taglinks;
+DROP TABLE globals;
 """
 
 ######## DATABASE CONNECTION AND SETUP ########
 
-def connect(app = None):
-    global conn, cur
+def connect(app_ = None):
+    global app, conn, cur
+    app = app_
     exists = os.path.exists('cdp.db')
     conn = sqlite3.connect('cdp.db', check_same_thread = False)
     cur = conn.cursor()
     if not exists:
         cur.executescript(SETUP)
         rebuild()
-    if app: add_url_rules(app)
+    if app:
+        add_url_rules(app)
 
 def rebuild():
-    if os.path.exists('music'):
-        os.rename('music', 'temp/music')
+    if os.path.exists('music') and os.listdir('music'):
+        try:
+            os.rename('music', 'temp/music')
+        except WindowsError:
+            time.sleep(.5)
+            os.rename('music', 'temp/music')
     tagdict = {}
     cur.execute("SELECT id, name, album_id, artist_id FROM songs")
     for id, name, album, artist in cur.fetchall():
         cur.execute("SELECT tag_id FROM taglinks WHERE song_id = ?", (id,))
         tags = [i[0] for i in cur]
         tagdict[name, album, artist] = tags
-    cur.executescript(RESET)
+    cur.executescript(CLEAR)
     if os.path.exists('temp/music'):
         errors = add_path('temp/music', delete = True)
         print 'errors: ' + str(errors)
+    update_fsrecords('music')
     for (name, album, artist), tagid in tagdict.iteritems():
         artist_id = Artist.get_id(artist)
         album_id = Album.get_id(album, artist_id)
@@ -124,10 +139,29 @@ def rebuild():
         if result:
             cur.execute("INSERT INTO taglinks (song_id, tag_id) VALUES (?,?)", (result[0], tagid))
     conn.commit()
-        
+
+def scan(wait_seconds = 0.1):
+    try:
+        _scan("music", wait_seconds)
+        cur.execute("SELECT path FROM fsrecords WHERE checked = 0 LIMIT 1")
+        row = cur.fetchone()
+        if row:
+            path = row[0]
+            node_type = "directory" if os.path.isdir(path) else "file"
+            raise FSChangedException("Missing %s: %s" % (node_type, path))
+        cur.execute("UPDATE fsrecords SET checked = 0")
+        conn.commit()
+    except FSChangedException as e:
+        print "[ERROR] " + e.message
+        print "Library files have been modified outside of this program!"
+        raw_input("[ENTER] to rebuild library (may fuck with http client, and will clear playlist)")
+        rebuild()
+
+# the following are helper functions and may make changes to the database
+# without commiting them.
+
 def add_path(path, delete = False, errors = []):
     if os.path.isdir(path):
-        update_folderhash(path)
         for node in os.listdir(path):
             node = os.path.join(path, node)
             errors += add_path(node, delete, errors)
@@ -136,27 +170,27 @@ def add_path(path, delete = False, errors = []):
     elif os.path.splitext(path)[1] in util.FORMATS:
 ##        try:
         id = add_song(*get_song_info(path))
-##        except Exception:
-##            errors += "duplicate",
+##        except: #duplicate files cause some sqlite exception
+##            errors += (path, "duplicate"),
 ##            return errors
         file_path, e = organize.moveFile(path, "music", delete)
         if e:
-            errors += (path, file_path, e)
+            Song.delete(id)
+            errors += (path, e)
             return errors
-        size = os.path.getsize(file_path)
-        cur.execute("UPDATE songs SET path = ?, size = ? WHERE id = ?",
-                    (file_path, size, id))
+        else:
+            Song.update(id, path = file_path)
     return errors
 
-def add_song(name, album, track_performer, artist):
+def add_song(name, album, track_performer, artist, size):
     cur.execute("INSERT OR IGNORE INTO artists (name) VALUES (?)", (artist,))
     cur.execute("SELECT id FROM artists WHERE name = ?", (artist,))
     artist_id = cur.fetchone()[0]
     cur.execute("INSERT OR IGNORE INTO albums (name, artist_id) VALUES (?, ?)", (album, artist_id))
     cur.execute("SELECT id FROM albums WHERE name = ?", (album,))
     album_id = cur.fetchone()[0]
-    cur.execute("INSERT INTO songs (name, album_id, track_performer, artist_id) VALUES (?,?,?,?)",
-                (name, album_id, track_performer, artist_id))
+    cur.execute("INSERT INTO songs (name, album_id, track_performer, artist_id, size) VALUES (?,?,?,?,?)",
+                (name, album_id, track_performer, artist_id, size))
     return cur.lastrowid
     
 def get_song_info(path):
@@ -165,7 +199,38 @@ def get_song_info(path):
     artist = tags['performer'] or track_performer
     album = tags['album']
     name = tags['title'] or os.path.splitext(os.path.split(path)[1])[0]
-    return (name, album, track_performer, artist)
+    return (name, album, track_performer, artist, os.path.getsize(path))
+
+def update_fsrecords(path):
+    update_fsrecord(path)
+    if os.path.isdir(path):
+        for node in os.listdir(path):
+            update_fsrecords(os.path.join(path, node))
+
+def update_fsrecord(path):
+    size = os.path.getsize(path) if os.path.isfile(path) else None
+    cur.execute("INSERT OR REPLACE INTO fsrecords VALUES (?, ?, ?)",
+                (path, size, None))
+
+def remove_fsrecord(path):
+    cur.execute("DELETE FROM fsrecords WHERE path = ?", (path,))
+
+class FSChangedException(Exception): pass
+def _scan(path, wait):
+    isdir = os.path.isdir(path)
+    cur.execute("SELECT * FROM fsrecords WHERE path = ? LIMIT 1", (path,))
+    row = cur.fetchone()
+    if not row:
+        node_type = "directory" if isdir else "file"
+        raise FSChangedException("Unexpected %s: %s" % (node_type, path))
+    cur.execute("UPDATE fsrecords SET checked = 1 WHERE path = ?", (path,))
+    if isdir:
+        for node in os.listdir(path):
+            if wait: time.sleep(wait)
+            _scan(os.path.join(path, node), wait)
+    else:
+        if row[1] != os.path.getsize(path):
+            raise FSChangedException("File changed: %s" % path)
 
 ######## CLASSES, GETTERS, AND UPDATERS ########
 
@@ -189,7 +254,7 @@ class Query():
     def artists(Cls, name = None):
         self = Cls()
         if name:
-            cur.execute("SELECT * FROM artists WHERE name = ?", name)
+            cur.execute("SELECT * FROM artists WHERE name = ?", (name,))
         else:
             cur.execute("SELECT * FROM artists")
         self.artists = [ShortArtist(*tup) for tup in cur.fetchall()]
@@ -255,6 +320,9 @@ class Song(Item):
         self.track_performer = track_performer
         self.path = path
         self.change_alerted = False
+
+    def __repr__(self):
+        return "Song(id = %s)" % self.id
         
     def check_fs(self):
         return os.path.exists(self.rel_path) \
@@ -283,14 +351,10 @@ class Song(Item):
         statement = "UPDATE songs SET"
         values = []
         for key, value in kwargs.iteritems():
-            set += " %s = ?" % key
+            statement += " %s = ?" % key
             values.append(value)
         values.append(id)
         cur.execute(statement + " WHERE id = ?", values)
-
-    @staticmethod
-    def delete(id):
-        cur.execute("DELETE FROM songs WHERE id = ?", (id,))
 
     @staticmethod
     def get_tags(id):
@@ -306,22 +370,30 @@ class Song(Item):
 
     @staticmethod
     def delete(id):
-        row = cur.execute("SELECT artist_id, album_id, path FROM songs WHERE id = ?", (id,))
+        cur.execute("SELECT artist_id, album_id, path FROM songs WHERE id = ?", (id,))
+        row = cur.fetchone()
         if not row: return
+        print "row == " + str(row)
         artist_id, album_id, path = row
-        update_tags(id)
+        Song.update_tags(id)
         cur.execute("DELETE FROM songs WHERE id = ?", (id,))
         #remove artist and album if no longer linked to by any songs
         cur.execute("SELECT EXISTS (SELECT 1 FROM songs WHERE album_id = ?)", (album_id,))
         if not cur.fetchone()[0]:
             cur.execute("DELETE FROM albums WHERE id = ?", (album_id,))
-            cur.execute("SELECT EXISTS (SELECT 1 FROM songs WHERE artist_id = ?)", (artist_id,))
-            if not cur.fetchone()[0]:
-                cur.execute("DELETE FROM albums WHERE id = ?", (artist_id,))
-        os.remove(path)
+        cur.execute("SELECT EXISTS (SELECT 1 FROM songs WHERE artist_id = ?)", (artist_id,))
+        if not cur.fetchone()[0]:
+            cur.execute("DELETE FROM artists WHERE id = ?", (artist_id,))
+        os.remove(path) #remove song
+        remove_fsrecord(path)
         path = os.path.split(path)[0]
-        if path and not os.listdir(path):
-            os.remove(path)
+        if path and not os.listdir(path) and path != 'music':
+            os.rmdir(path) #remove album folder
+            remove_fsrecord(path)
+            path = os.path.split(path)[0]
+            if path and not os.listdir(path) and path != 'music':
+                os.rmdir(path) #remove artist folder
+                remove_fsrecord(path)
         conn.commit()
 
 class Artist(Item):
@@ -349,7 +421,7 @@ class Artist(Item):
 
     @staticmethod
     def get_id(name):
-        cur.execute("SELECT id FROM artists WHERE name = ?", name)
+        cur.execute("SELECT id FROM artists WHERE name = ?", (name,))
         id = cur.fetchone()
         if id: return id[0]
     
@@ -390,67 +462,69 @@ class Album(Item):
         cur.execute("SELECT id FROM albums WHERE name = ?", (name,))
         return [i[0] for i in cur.fetchall()]
 
-def jsonify(data):
-    return Response(encoder.encode(data),
-                    mimetype = 'application/json')
-
-def JSON(function):
-    def func(*args, **kwargs):
-        data = function(*args, **kwargs)
-        if data == None:
-            abort(404)
-        resp = jsonify(data)
-        resp.status_code=200
-        return resp
-    func.__name__ = function.__name__
-    func.__doc__ = function.__doc__
-    return func
-
-def update_folderhash(path):
+def get_folderhash(path):
     mr_itchy = md5.md5()
     contents = os.listdir(path)
     for node in contents:
         if node not in IGNORE:
             mr_itchy.update(node)
-    hash = mr_itchy.hexdigest()
-    cur.execute("INSERT OR REPLACE INTO folderhashes VALUES (?, ?)", (path, hash))
+    return mr_itchy.hexdigest()
 
-@JSON
+def jsonify(data):
+    if data == None:
+        abort(404)
+    else:
+        return Response(encoder.encode(data),
+                        mimetype = 'application/json')
+
+routes = []
+
+def route(location, *methods, **kwargs):
+    def decorator(func):
+        global routes
+        if 'methods' not in kwargs and methods:
+            kwargs['methods'] = methods
+        routes.append(((location, func.__name__, func), kwargs))
+        return func
+    return decorator
+
+@route('/library')
 def get_artists():
-    return Query.artists()
+    return jsonify(Query.artists())
 
-@JSON
+@route('/library/artist')
 def get_artists_query():
     name = request.args.get("name")
-    return Query.artists(name)
+    return jsonify(Query.artists(name))
 
-@JSON
+@route('/library/artist/<int:id>')
 def get_artist(id):
-    return Artist.get(id)
+    return jsonify(Artist.get(id))
 
-@JSON
+@route('/library/album')
 def get_albums():
     name = request.args.get("name")
-    return Query.albums(name)
+    return jsonify(Query.albums(name))
 
-@JSON
+@route('/library/album/<int:id>')
 def get_album(id):
-    return Album.get(id)
+    return jsonify(Album.get(id))
 
 song_keys = ['name', 'artist', 'album', 'track_performer']
 
-@JSON
+@route('/library/song')
 def get_songs():
     restrictions = {}
     for key in song_keys:
         if request.args.get(key) != None:
             restrictions[key] = request.args.get(key)
-    return Query.songs(**restrictions)
+    return jsonify(Query.songs(**restrictions))
 
-@JSON
+@route('/library/song/<int:id>')
 def get_song(id):
-    return Song.get(id)
+    return jsonify(Song.get(id))
 
+@route('/library/song/<int:id>', 'PATCH')
 def update_song(id):
     changes = {}
     for key in song_keys:
@@ -458,70 +532,72 @@ def update_song(id):
     Song.update(id, **cull_dictionary(changes))
     return "200 OK"
 
+@route('/library/song/<int:id>', 'DELETE')
 def delete_song(id):
     Song.delete(id)
     return "200 OK"
 
-@JSON
+@route('/library/song/<int:id>/tags')
 def get_song_tags(id):
-    return Song.get_tags(id)
+    return jsonify(Song.get_tags(id))
 
+@route('/library/song/<int:id>/tags', 'PUT')
 def update_song_tags(id):
-    tags = request.args.get('tags').split(',')
+    cur.execute("SELECT id FROM tags")
+    valid_tags = [row[0] for row in cur]
+    tags = request.form['tags']
+    tags = tags.split(',')
     for i in xrange(len(tags)-1, -1, -1):
         try:
             tags[i] = int(tags[i])
         except:
             del tags[i]
+        if tags[i] not in valid_tags:
+            abort(400)
     Song.update_tags(id, tags)
     return "200 OK"
 
+@route('/library/song/<int:id>/tags', 'DELETE')
 def remove_song_tags(id):
     Song.update_tags(id)
     return "200 OK"
 
-@JSON
+@route('/library/tags')
 def get_tags():
     cur.execute("SELECT * FROM tags")
-    return [{'id':row[0],'name':row[1]} for row in cur]
+    return jsonify([{'id':row[0],'name':row[1]} for row in cur])
 
+@route('/library/tags', 'POST')
 def new_tag():
     name = request.form['name']
     cur.execute("INSERT OR IGNORE INTO tags (name) VALUES (?)", (name,))
     conn.commit()
-    return "200 OK"
+    return '200 OK'
 
+@route('/library/tags/<int:id>', 'DELETE')
 def del_tag(id):
     cur.execute("DELETE FROM taglinks WHERE tag_id = ?", (id,))
     cur.execute("DELETE FROM tags WHERE id = ?", (id,))
     conn.commit()
     return '200 OK'
 
-def add_url_rules(app):
-    app.add_url_rule('/library', 'get_artists', get_artists)
-    app.add_url_rule('/library/artist', 'get_artists_query', get_artists_query)
-    app.add_url_rule('/library/artist/<int:id>', 'get_artist', get_artist)
-    app.add_url_rule('/library/album', 'get_albums', get_albums)
-    app.add_url_rule('/library/album/<int:id>', 'get_album', get_album)
-    app.add_url_rule('/library/song', 'get_songs', get_songs)
-    app.add_url_rule('/library/song/<int:id>', 'get_song', get_song)
-    app.add_url_rule('/library/song/<int:id>', 'update_song', update_song,
-                     methods = ['PATCH'])
-    app.add_url_rule('/library/song/<int:id>', 'delete_song', delete_song,
-                     methods = ['DELETE'])
-    app.add_url_rule('/library/song/<int:id>/tags', 'get_song_tags',
-                     get_song_tags)
-    app.add_url_rule('/library/song/<int:id>/tags', 'update_song_tags',
-                     update_song_tags, methods = ['PUT'])
-    app.add_url_rule('/library/song/<int:id>/tags', 'remove_song_tags',
-                     remove_song_tags, methods = ['DELETE'])
-    app.add_url_rule('/library/tags', 'get_tags', get_tags)
-    app.add_url_rule('/library/tags', 'new_tags', new_tag, methods = ['POST'])
-    app.add_url_rule('/library/tags/<int:id>', 'del_tag', del_tag,
-                     methods = ['DELETE'])
+@route('/api/rebuild_database', 'POST')
+def rebuild_endpoint():
+    rebuild()
+    return '200 OK'
 
-if __name__ == '__main__':
+def add_url_rules(app):
+    for args, kwargs in routes:
+        app.add_url_rule(*args, **kwargs)
+
+def run():
     app = Flask(__name__)
     app.debug = DEBUG
     connect(app)
     app.run()
+
+if __name__ == '__main__':
+    if 'idlelib' not in sys.modules:
+        run()
+    else:
+        connect()
